+++ src/Transport/ExponentialBackoff.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 namespace Innmind\HttpTransport\Transport;
 
 use Innmind\HttpTransport\{
+    Config,
     Success,
     Information,
     Redirection,
@@ -80,10 +81,24 @@ final class ExponentialBackoff implements Implementation
     #[\Override]
     public function map(callable $map): self
     {
+        $fulfill = $this->fulfill->map($map);
+
         return self::of(
-            $this->fulfill->map($map),
-            $this->halt,
+            $fulfill,
+            $fulfill->config()->async()->match(
+                static fn($async) => $async->halt(),
+                fn() => $this->halt,
+            ),
         );
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
+    #[\Override]
+    public function config(): Config
+    {
+        return $this->fulfill->config();
     }
 
     /**