+++ src/Transport/Logger.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 namespace Innmind\HttpTransport\Transport;
 
 use Innmind\HttpTransport\{
+    Config,
     Success,
     Information,
     Redirection,
@@ -60,6 +61,15 @@ final class Logger implements Implementation
             $this->fulfill->map($map),
             $this->logger,
         );
+    }
+
+    /**
+     * @psalm-mutation-free
+     */
+    #[\Override]
+    public function config(): Config
+    {
+        return $this->fulfill->config();
     }
 
     private function logRequest(Request $request): string